updates for serde

+++ src/ulamspiral_img.rs
@@ -4,7 +4,7 @@ pub type GreyImage = image::GrayImage; // because ben is british
 
 pub fn generate(x_size: u32, y_size: u32) -> Result<GreyImage, Box<dyn Error>> {
     let mut img = image::ImageBuffer::new(x_size, y_size);
-    let total = u32::from(cmp::max(x_size, y_size)).pow(2).try_into()?;
+    let total = cmp::max(x_size, y_size).pow(2).try_into()?;
     let sieve = primal::Sieve::new(total);
     let pixel = image::Luma::from([255]);
 
@@ -36,7 +36,7 @@ pub fn generate_colour(
     y_size: u32,
 ) -> Result<image::RgbImage, Box<dyn Error>> {
     let mut img = image::RgbImage::new(x_size, y_size);
-    let total = u32::from(cmp::max(x_size, y_size)).pow(2).try_into()?;
+    let total = cmp::max(x_size, y_size).pow(2).try_into()?;
     let sieve = primal::Sieve::new(total);
 
     for prime in sieve.primes_from(0).take_while(|x| *x <= total) {
@@ -73,25 +73,25 @@ mod tests {
     fn check_sq() {
         let odd = generate(5, 5).unwrap();
         let mut expected_odd = image::ImageBuffer::new(5, 5);
-        expected_odd.put_pixel(3, 0, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(0, 1, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(2, 1, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(3, 1, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(4, 1, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(1, 2, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(0, 3, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(2, 3, image::Luma::from([255 as u8]));
-        expected_odd.put_pixel(3, 4, image::Luma::from([255 as u8]));
+        expected_odd.put_pixel(3, 0, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(0, 1, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(2, 1, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(3, 1, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(4, 1, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(1, 2, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(0, 3, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(2, 3, image::Luma::from([255_u8]));
+        expected_odd.put_pixel(3, 4, image::Luma::from([255_u8]));
         assert_eq!(odd, expected_odd);
 
         let even = generate(4, 4).unwrap();
         let mut expected_even = image::ImageBuffer::new(4, 4);
-        expected_even.put_pixel(2, 0, image::Luma::from([255 as u8]));
-        expected_even.put_pixel(1, 1, image::Luma::from([255 as u8]));
-        expected_even.put_pixel(2, 1, image::Luma::from([255 as u8]));
-        expected_even.put_pixel(3, 1, image::Luma::from([255 as u8]));
-        expected_even.put_pixel(0, 2, image::Luma::from([255 as u8]));
-        expected_even.put_pixel(1, 3, image::Luma::from([255 as u8]));
+        expected_even.put_pixel(2, 0, image::Luma::from([255_u8]));
+        expected_even.put_pixel(1, 1, image::Luma::from([255_u8]));
+        expected_even.put_pixel(2, 1, image::Luma::from([255_u8]));
+        expected_even.put_pixel(3, 1, image::Luma::from([255_u8]));
+        expected_even.put_pixel(0, 2, image::Luma::from([255_u8]));
+        expected_even.put_pixel(1, 3, image::Luma::from([255_u8]));
         assert_eq!(even, expected_even);
     }
     #[test]
@@ -112,23 +112,23 @@ mod tests {
     fn check_rect() {
         let tall = generate(5, 3).unwrap();
         let mut expected_tall = image::ImageBuffer::new(5, 3);
-        expected_tall.put_pixel(0, 0, image::Luma::from([255 as u8]));
-        expected_tall.put_pixel(2, 0, image::Luma::from([255 as u8]));
-        expected_tall.put_pixel(3, 0, image::Luma::from([255 as u8]));
-        expected_tall.put_pixel(4, 0, image::Luma::from([255 as u8]));
-        expected_tall.put_pixel(1, 1, image::Luma::from([255 as u8]));
-        expected_tall.put_pixel(0, 2, image::Luma::from([255 as u8]));
-        expected_tall.put_pixel(2, 2, image::Luma::from([255 as u8]));
+        expected_tall.put_pixel(0, 0, image::Luma::from([255_u8]));
+        expected_tall.put_pixel(2, 0, image::Luma::from([255_u8]));
+        expected_tall.put_pixel(3, 0, image::Luma::from([255_u8]));
+        expected_tall.put_pixel(4, 0, image::Luma::from([255_u8]));
+        expected_tall.put_pixel(1, 1, image::Luma::from([255_u8]));
+        expected_tall.put_pixel(0, 2, image::Luma::from([255_u8]));
+        expected_tall.put_pixel(2, 2, image::Luma::from([255_u8]));
         assert_eq!(tall, expected_tall);
 
         let wide = generate(3, 5).unwrap();
         let mut expected_wide = image::ImageBuffer::new(3, 5);
-        expected_wide.put_pixel(2, 0, image::Luma::from([255 as u8]));
-        expected_wide.put_pixel(1, 1, image::Luma::from([255 as u8]));
-        expected_wide.put_pixel(2, 1, image::Luma::from([255 as u8]));
-        expected_wide.put_pixel(0, 2, image::Luma::from([255 as u8]));
-        expected_wide.put_pixel(1, 3, image::Luma::from([255 as u8]));
-        expected_wide.put_pixel(2, 4, image::Luma::from([255 as u8]));
+        expected_wide.put_pixel(2, 0, image::Luma::from([255_u8]));
+        expected_wide.put_pixel(1, 1, image::Luma::from([255_u8]));
+        expected_wide.put_pixel(2, 1, image::Luma::from([255_u8]));
+        expected_wide.put_pixel(0, 2, image::Luma::from([255_u8]));
+        expected_wide.put_pixel(1, 3, image::Luma::from([255_u8]));
+        expected_wide.put_pixel(2, 4, image::Luma::from([255_u8]));
         assert_eq!(wide, expected_wide);
     }
 }
